Fix project entity test
